TS: Factor out util types into their own package
