+++ nix/jl4-web/package.nix
@@ -30,6 +30,10 @@ buildNpmPackage rec {
 
     pushd ./ts-shared
 
+    pushd ./type-utils
+    npm run build
+    popd
+
     pushd ./viz-expr
     npm run build
     popd
