+++ ts-shared/l4-ladder-visualizer/package.json
@@ -73,6 +73,7 @@
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
     "@repo/layout-ir": "*",
+    "@repo/type-utils": "*",
     "@repo/viz-expr": "*",
     "@tailwindcss/cli": "^4.0.9",
     "@tailwindcss/vite": "^4.0.6",
