+++ ts-shared/l4-ladder-visualizer/src/lib/algebraic-graphs/adjacency-map-directed-graph.ts
@@ -1,5 +1,5 @@
 import { BaseAMGraph } from './base-adjacency-map.js'
-import type { Eq, Ord } from '@repo/layout-ir'
+import type { Eq, Ord } from '@repo/type-utils'
 import {
   DirectedEdge,
   stringifyEdge as makeEdgeKey,
