+++ ts-shared/l4-ladder-visualizer/src/lib/algebraic-graphs/dag.ts
@@ -1,4 +1,4 @@
-import type { Eq, Ord, HasId } from '@repo/layout-ir'
+import type { Eq, Ord, HasId } from '@repo/type-utils'
 import {
   DirectedAMGraph,
   makeDirectedConnectAdjacencyMap,
