+++ ts-shared/l4-ladder-visualizer/src/lib/utils.ts
@@ -4,17 +4,6 @@ import type { Properties as CSSProperties } from 'csstype'
  *********** Types and type combinators **********
  *************************************************/
 
-/*************************
-      Branded types
-**************************/
-
-/** https://egghead.io/blog/using-branded-types-in-typescript
- *  See also the TS playground excerpt linked from the article above
- */
-declare const __brand: unique symbol
-type Brand<B> = { [__brand]: B }
-export type Branded<T, B> = T & Brand<B>
-
 /* ====== CSS ============ */
 
 export type Styles = {
