+++ ts-shared/layout-ir/index.ts
@@ -1,2 +1,1 @@
 export * from './core.js'
-export * from './utils.js'
